requirements.txt
asgiref==3.2.10
astroid==2.4.2
attrs==19.3.0
backcall==0.2.0
decorator==4.4.2
Django==3.1
django-rest-framework==0.1.0
djangorestframework==3.11.1
extras==1.0.0
factory-boy==2.12.0
Faker==4.1.1
fixtures==3.0.0
importlib-metadata==1.7.0
iniconfig==1.0.1
ipython==7.16.1
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.17.2
lazy-object-proxy==1.4.3
linecache2==1.0.0
logilab-common==1.7.2
mccabe==0.6.1
more-itertools==8.4.0
mypy-extensions==0.4.3
packaging==20.4
parso==0.7.1
pbr==5.4.5
pexpect==4.8.0
pickleshare==0.7.5
pkg-resources==0.0.0
pluggy==0.13.1
prompt-toolkit==3.0.5
ptyprocess==0.6.0
py==1.9.0
Pygments==2.6.1
pylint==2.5.3
pylint-django==2.3.0
pylint-plugin-utils==0.6
pyparsing==2.4.7
pytest==6.0.1
pytest-django==3.9.0
python-dateutil==2.8.1
python-mimeparse==1.6.0
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
testtools==2.4.0
text-unidecode==1.3
toml==0.10.1
traceback2==1.4.0
traitlets==4.3.3
typed-ast==1.4.1
typing-extensions==3.7.4.2
unittest2==1.1.0
wcwidth==0.2.5
wrapt==1.12.1
zipp==3.1.0
